pyproject.toml
[project]
name = "first-evolution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.105.0",
    "crewai-tools>=0.38.1",
    "docling>=2.26.0",
    "evolutionapi>=0.1.1",
    "flask>=3.1.0",
    "litellm>=1.60.2",
    "pdf2image>=1.17.0",
    "pypandoc>=1.15",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
